server: split http handling and functionality (part2)
